requirements.txt
Django
pytz==2024.2
Pillow==10.4.0
asgiref==3.8.1
sqlparse==0.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-redis==5.4.0
django-cryptography==1.1
drf-spectacular==0.27.2
drf-yasg==1.21.8
psycopg2-binary==2.9.9
python-dotenv==1.0.1
social-auth-app-django==5.4.2
pytz==2024.2
waitress==3.0.0
daphne==4.1.2
celery==5.4.0
redis==5.1.1
numpy==2.0.2
firebase-admin==6.6.0
django-schema-history==0.1.0
django-cors-headers==4.4.0
gunicorn==23.0.0
google-auth==2.28.2
google-auth-oauthlib==1.2.0
requests==2.31.0
google-auth-httplib2==0.2.0
django-elasticsearch-dsl==8.0.0
elasticsearch
supabase==2.13.0
channels==4.0.0
channels_redis==4.1.0
uvicorn
whitenoise>=6.2
django_extensions